+++ src/styles/components/ui/Typography/index.ts
@@ -1,0 +1,2 @@
+import Typography from "./Typography";
+export default Typography;
